Initial commit for demo.
I got the infinite scroll directive to be more stable. I did
this by switching from jQuery to Zepto and commenting out the assignment
to '$' -- since that would typically conflict with any framework=
that the instrumented page is trying to use. I also moved the
javascript-in-a-string code to its own file so that it's easier to
proofread and debug.

Lastly, added demo.py which is a little driver for running a single
directive on a single URL.