Trying to debug infinite scrolling.
Scrolling has some sensitivity to being called too quickly. It is more
reliable if we impose a delay in between each call to scrollTo. In order
to implement this and still be able to wait for XHRs to finish, I added
a _very_ lightweight promises system. In keeping with the spirit of POJS,
I opted not to use a promises framework. (The version of webkit in splash
does not have native promises, either.)