Finished the overlay directive.
This is the last of the directives. The former overlay detection
used hard-coded CSS selectors that only worked on Twitter and
dontsellbodies.org. I've improved the heuristic a bit here so
that it will generalize better.

I also cleaned up all of the examples in this commit. This concludes
the main work on the kitchen sink directive. Next, I will be testing
the new directives against some real sites.